fix(app): refactoring

+++ apps/tg-bot/src/app/modules/observatory/services/observatory.service.ts
@@ -1,14 +1,17 @@
-import {Inject, Injectable, OnModuleInit} from '@nestjs/common';
+import {Inject, Injectable, OnApplicationBootstrap, OnModuleInit} from '@nestjs/common';
 import {Menu} from '@grammyjs/menu';
 import {BotContext} from '../../bot/interfaces/bot-context.interface';
-import {MemeModerationMenusEnum} from '../../conversations/constants/meme-moderation-menus.enum';
 import {UserPermissionEnum} from '../../bot/constants/user-permission.enum';
 import {UserService} from '../../bot/services/user.service';
-import {Bot, InlineKeyboard} from "grammy";
-import {BOT} from "../../bot/providers/bot.provider";
-import {BaseConfigService} from "../../config/base-config.service";
-import {UserRequestService} from "../../bot/services/user-request.service";
-import {ClientBaseService} from "../../client/services/client-base.service";
+import {Bot, InlineKeyboard} from 'grammy';
+import {BOT} from '../../bot/providers/bot.provider';
+import {BaseConfigService} from '../../config/base-config.service';
+import {UserRequestService} from '../../bot/services/user-request.service';
+import {ClientBaseService} from '../../client/services/client-base.service';
+import {ObservatoryPostMenusEnum} from '../contsants/observatory-post-menus.enum';
+import {InjectRepository} from '@nestjs/typeorm';
+import {Repository} from 'typeorm';
+import {ObservatoryPostEntity} from '../entities/observatory-post.entity';
 
 @Injectable()
 export class ObservatoryService implements OnModuleInit {
@@ -17,54 +20,60 @@ export class ObservatoryService implements OnModuleInit {
     private baseConfigService: BaseConfigService,
     private userService: UserService,
     private userRequestService: UserRequestService,
-    private clientBaseService: ClientBaseService
+    private clientBaseService: ClientBaseService,
+    @InjectRepository(ObservatoryPostEntity)
+    private observatoryPostRepository: Repository<ObservatoryPostEntity>
   ) {
   }
 
   /**
    * Меню публикации одобренного поста
    */
-  private observerPostMenu: Menu<BotContext>;
+  private observatoryPostMenu: Menu<BotContext>;
 
   public onModuleInit(): void {
+    this.onNewObservatoryPost();
+    this.waitDeleteObserverPost();
     this.buildObservatoryPostMenu();
-    this.onObserverStationPost();
   }
 
-
-  private onObserverStationPost() {
+  private onNewObservatoryPost() {
     this.clientBaseService.observerChannelPost$.subscribe(async (ctx) => {
-      await ctx.api.copyMessage(
+      const message = await ctx.api.copyMessage(
         this.baseConfigService.userRequestMemeChannel,
         ctx.channelPost.sender_chat.id,
         ctx.channelPost.message_id,
-        {disable_notification: true, caption: '', reply_markup: this.observerPostMenu}
+        {disable_notification: true, caption: '', reply_markup: this.observatoryPostMenu}
       );
+
+      const post = await this.observatoryPostRepository.create({
+        requestChannelMessageId: message.message_id,
+      });
+      await this.observatoryPostRepository.save(post);
     });
   }
 
   private buildObservatoryPostMenu(): void {
-    this.observerPostMenu = new Menu<BotContext>(MemeModerationMenusEnum.OBSERVATORY_POST, {
+    this.observatoryPostMenu = new Menu<BotContext>(ObservatoryPostMenusEnum.POST_MENU, {
       autoAnswer: false,
     })
       .text('🤖 Пост обсерватории')
       .row()
       .text('Отклонить', async (ctx) => {
-        if (this.userService.checkPermission(ctx, UserPermissionEnum.ALLOW_DELETE_REJECTED_POST)) {
+        if (this.userService.checkPermission(ctx, UserPermissionEnum.IS_BASE_MODERATOR)) {
           await this.rejectObserverPost(ctx);
         }
       })
       .text('Опубликовать', async (ctx) => {
-        if (this.userService.checkPermission(ctx, UserPermissionEnum.ALLOW_PUBLISH_TO_CHANNEL)) {
-          await this.publishObserverPost(ctx);
+        if (this.userService.checkPermission(ctx, UserPermissionEnum.IS_BASE_MODERATOR)) {
+          await this.publishObservatoryPost(ctx);
         }
       })
       .row();
-    this.bot.use(this.observerPostMenu);
+    this.bot.use(this.observatoryPostMenu);
   }
 
-
-  private async publishObserverPost(ctx: BotContext): Promise<void> {
+  private async publishObservatoryPost(ctx: BotContext): Promise<void> {
     const channelInfo = await ctx.api.getChat(this.baseConfigService.memeChanelId);
     const link = channelInfo['username']
       ? `https://t.me/${channelInfo['username']}`
@@ -82,15 +91,48 @@ export class ObservatoryService implements OnModuleInit {
       }
     );
 
+    await this.observatoryPostRepository.update(
+      {requestChannelMessageId: ctx.callbackQuery.message.message_id},
+      {
+        publishedMessageId: publishedMessage.message_id,
+        isApproved: true,
+        processedByModerator: {id: ctx.callbackQuery.from.id},
+      }
+    );
+
     const postLink = channelInfo['username']
       ? `https://t.me/${channelInfo['username']}/${publishedMessage.message_id}`
       : channelInfo['invite_link'];
 
-    const inlineKeyboard = new InlineKeyboard().url('🤖 Опубликован', postLink).row();
+    const inlineKeyboard = new InlineKeyboard()
+      .url(`🤖 Опубликован (${ctx.callbackQuery.from.username})`, postLink)
+      .row();
     await ctx.editMessageReplyMarkup({reply_markup: inlineKeyboard});
   }
 
   private async rejectObserverPost(ctx: BotContext): Promise<void> {
-    await ctx.deleteMessage();
+    await this.observatoryPostRepository.update(
+      {requestChannelMessageId: ctx.callbackQuery.message.message_id},
+      {
+        isApproved: false,
+        processedByModerator: {id: ctx.callbackQuery.from.id},
+      }
+    );
+
+    const inlineKeyboard = new InlineKeyboard()
+      .text(
+        `🤖 Отклонен ❌ (${ctx.callbackQuery.from.username})`,
+        ObservatoryPostMenusEnum.DELETE_OBSERVER_POST
+      )
+      .row();
+    await ctx.editMessageReplyMarkup({reply_markup: inlineKeyboard});
+  }
+
+  private waitDeleteObserverPost(): void {
+    this.bot.callbackQuery(ObservatoryPostMenusEnum.DELETE_OBSERVER_POST, async (ctx) => {
+      if (this.userService.checkPermission(ctx, UserPermissionEnum.ALLOW_DELETE_REJECTED_POST)) {
+        await ctx.deleteMessage();
+      }
+    });
   }
 }
